requirements.txt
coverage==4.5.1
Faker==0.8.17
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.11.0
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
nose==1.3.7
psycopg2-binary
python-dotenv==0.9.1
six
